Add AppWindowControl
