+++ frontend/src/app/window/control.tsx
@@ -1,0 +1,39 @@
+import { appWindow } from '@tauri-apps/api/window';
+import { ControlButtons, ControlType, WindowControl } from '../../components/window/control';
+
+export type AppWindowControlProp = {
+  buttons?: ControlButtons,
+
+  className?: string
+}
+
+export const AppWindowControl = ({
+  buttons,
+
+  className,
+}: AppWindowControlProp) => {
+  function onControlClick(type: ControlType) {
+    switch (type) {
+      case 'close': {
+        appWindow.close().then();
+        break;
+      }
+
+      case 'maximize': {
+        appWindow.toggleMaximize().then();
+        break;
+      }
+
+      case 'minimize': {
+        appWindow.minimize().then();
+        break;
+      }
+    }
+  }
+
+  return <WindowControl
+    className={className}
+    buttons={buttons}
+    onControlClick={onControlClick}
+  />;
+};
